+++ Snabble/Core/Products/ProductDB+Network.swift
@@ -8,17 +8,23 @@ import Foundation
 
 enum AppDbResponse {
     case diff(lines: String)
-    case full(db: Data, revision: Int)
+    case full(db: Data)
     case noUpdate   // no data, used when we get http 304
     case httpError  // http or network error
     case dataError  // data error: invalid content-type, unparsable data or other weird things
+    case aborted    // download was aborted. app may attempt resuming later
 }
 
 extension ProductDB {
 
-    private static let sqlType = "application/vnd+snabble.appdb+sql"
-    private static let sqliteType = "application/vnd+snabble.appdb+sqlite3"
+    fileprivate static let sqlType = "application/vnd+snabble.appdb+sql"
+    fileprivate static let sqliteType = "application/vnd+snabble.appdb+sqlite3"
     private static let contentTypes = "\(sqlType),\(sqliteType)"
+
+    func appDbSession(_ completion: @escaping (AppDbResponse) -> ()) -> URLSession {
+        let delegate = AppDBDownloadDelegate(self, completion)
+        return URLSession(configuration: .default, delegate: delegate, delegateQueue: nil)
+    }
 
     func getAppDb(currentRevision: Int64, schemaVersion: String, forceFullDb: Bool = false, completion: @escaping (AppDbResponse) -> () ) {
         let parameters = [
@@ -26,60 +32,118 @@ extension ProductDB {
             "schemaVersion": schemaVersion
         ]
 
+        if self.downloadTask != nil {
+            Log.warn("appDB download task already running, ignoring update request")
+            return
+        }
+
         self.project.request(.get, self.project.links.appdb.href, json: false, parameters: parameters, timeout: 0) { request in
             guard var request = request else {
                 return completion(.httpError)
             }
 
             request.setValue(ProductDB.contentTypes, forHTTPHeaderField: "Accept")
-            let session = SnabbleAPI.urlSession()
-            let start = Date.timeIntervalSinceReferenceDate
-            let task = session.dataTask(with: request) { data, response, error in
-                let elapsed = Date.timeIntervalSinceReferenceDate - start
-                let url = request.url?.absoluteString ?? "n/a"
-                Log.info("get \(url) took \(elapsed)s")
 
-                if error != nil {
-                    completion(.httpError)
-                    return
-                }
-
-                if let data = data, let response = response as? HTTPURLResponse {
-                    if response.statusCode == 304 {
-                        completion(.noUpdate)
-                        return
-                    }
-
-                    let headers = response.allHeaderFields
-                    if let contentType = headers["Content-Type"] as? String {
-                        if contentType == ProductDB.sqliteType {
-                            if let etag = headers["Etag"] as? String {
-                                completion(.full(db: data, revision: self.parseEtag(etag)))
-                                return
-                            }
-                        } else if contentType == ProductDB.sqlType {
-                            if let str = String(bytes: data, encoding: .utf8) {
-                                completion(.diff(lines: str))
-                                return
-                            }
-                        }
-                    }
-
-                    completion(.dataError)
-                }
-            }
+            let session = self.appDbSession(completion)
+            let task = session.downloadTask(with: request)
             task.resume()
+            self.downloadTask = task
         }
     }
 
-    // parse an ETag header value to extract the db revision
-    // format is either W/"xyz" (for weak tags) or simply "xyz"
-    private func parseEtag(_ etag: String) -> Int {
-        let isWeak = etag.lowercased().hasPrefix("w/")
-        let startIndex = isWeak ? etag.index(etag.startIndex, offsetBy: 2) : etag.startIndex
-        let tagValue = etag[startIndex...]
-        let stripped = tagValue.trimmingCharacters(in: CharacterSet(charactersIn: "\""))
-        return Int(stripped) ?? 0
+    func resumeAppDbDownload(_ completion: @escaping (AppDbResponse) -> () ) {
+        guard let resumeData = self.resumeData else {
+            return
+        }
+
+        if self.downloadTask != nil {
+            Log.warn("appDB download task already running, ignoring resume request")
+            return
+        }
+
+        Log.info("resuming download of appdb")
+        
+        let session = self.appDbSession(completion)
+        let task = session.downloadTask(withResumeData: resumeData)
+        task.resume()
+        self.downloadTask = task
     }
 
+}
+
+class AppDBDownloadDelegate: CertificatePinningDelegate, URLSessionDownloadDelegate {
+
+    private var completion: (AppDbResponse) -> ()
+    private var response: URLResponse?
+    private weak var productDb: ProductDB?
+    private let start = Date.timeIntervalSinceReferenceDate
+    private var bytesReceived: Int64 = 0
+    private var mbps = 0.0 // megabytes/second
+
+    init(_ productDb: ProductDB, _ completion: @escaping (AppDbResponse) -> ()) {
+        self.productDb = productDb
+        self.completion = completion
+    }
+
+    // MARK: - download delegate
+
+    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didWriteData bytesWritten: Int64, totalBytesWritten: Int64, totalBytesExpectedToWrite: Int64) {
+        let elapsed = Date.timeIntervalSinceReferenceDate - self.start
+        self.bytesReceived += bytesWritten
+
+        if elapsed > 0 {
+            self.mbps = Double(self.bytesReceived) / elapsed / 1024 / 1024
+        }
+        // print("download progress: \(totalBytesWritten ) \(self.mbps)")
+    }
+    
+    func urlSession(_ session: URLSession, downloadTask: URLSessionDownloadTask, didFinishDownloadingTo location: URL) {
+        let url = downloadTask.currentRequest?.url?.absoluteString ?? "n/a"
+        let elapsed = Date.timeIntervalSinceReferenceDate - self.start
+        Log.info("get \(url) took \(elapsed)s")
+
+        self.productDb?.downloadTask = nil
+        let fileData = try? Data(contentsOf: location)
+        if let data = fileData, let response = downloadTask.response as? HTTPURLResponse {
+            // print("got bytes: \(data.count) \(self.bytesReceived), \(self.mbps) MB/s")
+            if response.statusCode == 304 {
+                completion(.noUpdate)
+                return
+            }
+
+            let headers = response.allHeaderFields
+            if let contentType = headers["Content-Type"] as? String {
+                if contentType == ProductDB.sqliteType {
+                    completion(.full(db: data))
+                    return
+                } else if contentType == ProductDB.sqlType {
+                    if let str = String(bytes: data, encoding: .utf8) {
+                        completion(.diff(lines: str))
+                        return
+                    }
+                }
+            }
+
+            completion(.dataError)
+        }
+    }
+
+    func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
+        self.productDb?.resumeData = nil
+        self.productDb?.downloadTask = nil
+        guard let error = error as NSError? else {
+            return
+        }
+
+        let url = task.currentRequest?.url?.absoluteString ?? "n/a"
+        Log.info("\(url) finished with error \(error.code)")
+
+        let userInfo = error.userInfo
+        if let resumeData = userInfo[NSURLSessionDownloadTaskResumeData] as? Data {
+            self.productDb?.resumeData = resumeData
+            self.completion(.aborted)
+        } else {
+            self.completion(.httpError)
+        }
+    }
 }