+++ Snabble/UI/Payment/SepaCheckoutViewController.swift
@@ -14,7 +14,7 @@ final class SepaCheckoutViewController: UIViewController {
     @IBOutlet weak var cancelButton: UIButton!
     @IBOutlet weak var codeWidth: NSLayoutConstraint!
     
-    private weak var cart: ShoppingCart!
+    private var cart: ShoppingCart
     private weak var delegate: PaymentDelegate!
 
     private var process: CheckoutProcess
@@ -51,6 +51,8 @@ final class SepaCheckoutViewController: UIViewController {
         self.codeWidth.constant = self.codeImage.image?.size.width ?? 0
 
         self.cancelButton.setTitle("Snabble.Cancel".localized(), for: .normal)
+
+        self.navigationItem.hidesBackButton = true
     }
 
     override func viewWillAppear(_ animated: Bool) {